symmetric ref obj might be deleted before ref is deleted
Description: Symmetric refs may be created between similar type
objects. Their creation and deletion is independent of other.
When an object is being deleted, its link in referred objects are
to be deleted. But, it could so happen, that the referred object
also might have deleted while this object is being deleted.
In which case the ref object might not be available. Once link in
ref object is deleted, timestamps also also updated. So moved
the deletion of link and updation of time stamp under try for
safer access.

Change-Id: Iccf0d932058124aa4a64ff826c27107f41d455a2
Closes-Bug: #1729088