Optimize rule evaluation during flow setup.
When out network policy has evaluated action as drop for the flow, there is no
need to further evaluate SG or Firewall rules as the action for the flow will
remain drop regardless of the result of evaluation of SG and firewall rules.

Also update Flow introspect to show application-policy-set action summary.

Change-Id: I020019e674becdaf0dc44cdc93a57a7b69126cb4
Closes-Bug: #1730989