User configured Route target not seen in control-node
Root cause:
When we add a reference(uuid1), delete it and add same FQ name reference(uuid2) again.
We do not remove first reference from cache. In cache, the key is related to UUID. So
the above operation result is two items which has same FQ name exist in cache, because
the first one refreshed is false, in fact, we will send two message to ifmap:
          Add “to FQ name, uuid uuid1” reference
          Deleted “to FQ name, uuid uuid2” reference
From configure management client to ifmap we only use “to FQ name”, Add/Delete pair
is considered as same one.

Change-Id: I796499ad92b9748d40003a107da078c1aea794f0
Closes-bug: 1730363