Handle change of HC type
HC type was changed from segment to link-local. In this case, the earlier
segment instance would be deleted and a new one will be created for link local.
However the destroy was invoked after the object was modified and hence
assert of type was failing. Removing the assert as this is not required.

Change-Id: I47c40723961b540f9db4d4497731d21980af2b0f
closes-bug: #1732097