Process HC config prior to VMI
BGPaaS object is not using config manager (it has to inherit from
OperIFMapTable). Due to this, reevaluate of BGPaaS upon change of
HC object cannot be triggered via dependency manager. When VMI
update triggers BGPaaS config evaluation, which in turn evaluates
BFD health check, it is possible that HC entry is not created. Lack
of a further trigger could lead to HC not being initiated.

Re-ordering the config procesing such that HC is processed prior to VMI.

Change-Id: I374f985f043ed7a7f610b2063bac74f74a5562c2
closes-bug: #1732121