Check for health check in Flow Management
Flow entry was updated with HC uuid and is enqueued to flow management.
HC update was also enqueued to flow management. In case flow entry doesnt
get HC uuid (if HC is not available by then), but HC update was
processed prior to flow entry in flow management, HC is not started on the
flow. Fix this by checking for HC in flow management itself (removing HC
info in flow).

Change-Id: I7cc0b383acfd499fa48fec56637e39647d006286
closes-bug: #1733606