Do not delete XmppConnection if there is a reference (due to GR Helper)
During GR helper mode, BgpXmpppChannel hold all the routes as stale. In this
phase, if xmpp conenction is deleted for any reason, such as configuration
change, or control-node shutdown, then the GR needs to be aborted and all
routes need to be cleaned up. During this process, connection should not be
deleted yet. This reference check was not considered as part of MayDelete()
in XmppServerConnection cleanup.

Also make sure that XmppChannelMux::UnRegisterWriteReady() is called from all
clients to ensure proper clean up.

Unit test shall be added in a subsequent commit to cover this and other related
scenarios wrt GR configuration changes.

Change-Id: I472812a6dd5e52397d3723b5e7f99b320dae2728
Partial-Bug: 1733981