Policy of VlanNH should be based on VMI’s policy status
Currently VlanNH is always created with policy disabled NH. This will result in flows not getting created on transparent SI VMIs even when it has policy enabled. This resulted in a scenario where flows were getting created in reply path instead of request path. Because of this reply path traffic was considered as forward traffic/initiator of traffic which resulted in incorrect VNs being reported in Sessionlog messages.

Change-Id: Ia5f913bb599faf9c1ec2c75ea916c68aed1162aa
Closes-Bug: #1728802