Fix for analytics-api UVE query response delay
For every UVE request, analytics-api creates a connection to redis to
read the UVE. Creating a new redis connection for every UVE request is an
overhead and the response time depends on the gevent scheduling.
With this patch, analytics-api shares the same redis connections across the
UVE requests and therefore ensures the ordering of the UVE response per
redis, thereby alleviating the disproportionate response time for some
UVE requests.

Change-Id: Ibcc2306021a65f82da5d8898fb0d3fc7ab8c3c03
Closes-Bug: #1731187