* Pick right ksync queue for processing ksync entry.
Route ksync object would be deleted when all the route ksync entry are
deleted, deletion of ksync object happens in KSync task. KSync entry should
added and deleted in KSync context, but due to bug KSync entry was
getting deleted in UVE context resulting in immediate delete of ksync
object which might be in active use. Correcting the same.

Change-Id: Iac22cafe55e6bf240c2de0f06946349474b8c751
Closes-bug: #1727661