Remove kube-manager extra delete namespace events
kube-manager on a set timer tries to sync its local project object
cache with kubernetes namespaces and generates a delete namespace
event to clean up the project object if the associated k8s namespace
does not exist in the cache. But never removes the association so
the events keep getting generated.

Change-Id: I34190ceb9b1b464a94e12e514fff4aac16df7ed2
Closes-Bug: #1738788