Check for length of fq name array
While converting fq name to uuid, we are currently not checking that the length
of the fq name passed is same as the one found. As a result, a prefix is returned
as success. E.g. if we search for ['a', 'b'], but we found ['a', 'b', 'c'], we
would return success. This is not correct. We should raise NoIdError in such
cases.

Change-Id: Idc211ddcd31d4fd12ea4aba5179a7dc16c0c4375
Closes-Bug: 1729655