Added way to find vmi_uuid if it is not given in the annotation
When loadbalancer member is created by neutron, there is no
way to give the vmi uuid as annotation in the loadbalancer member
object. Since it is possible to find the vmi uuid from instance ip,
the loadbalancer driver will find the vmi uuid from instance_ip_address,
if the vmi annotation is not given for loadbalancer member object.

Change-Id: Ia6fa8eb2defcbe9f714518d9d1893124e733e29f
Partial-fix: #1740996