Update DNS Server List to connect to other servers.
On detecting current DNS Server connectivity down after a retry period, we
were obtaining the XMPP Server list and updating the list instead. This
could potentially result in connecting to the same DOWN server and not
updating the DNS records to the right bind server, resulting in DNS record
resolution failure.

The DNS server list would be updated at a random time as a result of subscribe
request to Discovery Server resulting in connection to the UP DNS server when
the records would be updated and record resolution succeeds.

Agent connects to the first two DNS server in the list.

It takes about 30 secs to connect to newer DNS server due to exponenential
retries to the older server. In addition, time to update the records to the
new bind server before record resolution succeeds.

Change-Id: If918b2726dcff1c5d3db9965bc6b36958cab5cb8
Partial-Bug: #1738304