Fix for analytics-api UVE query response delay
For every UVE request, analytics-api creates a connection to redis to
read the UVE. Creating a new redis connection for every UVE request is an
overhead and the response time depends on the gevent scheduling.
With this patch, analytics-api shares the same redis connections across the
UVE requests and therefore ensures the ordering of the UVE response per
redis, thereby alleviating the disproportionate response time for some
UVE requests.

Change-Id: I9b0b360b7ffc88c7524b9ae56bfde745468893e8
Closes-Bug: #1737486