Do not bring up peers when GR process is under progress
Peer sessions are not allowed to come up when GR is in progress. Once the state
enters GR_TIMER_WAIT or LLGR_TIMER_WAIT, then the sessions are allowed to come
up. However, it can so happen that when the peers are about to come up and
become established, the GR timers fire.. In this case, we do not want the peers
to come up to keep the design simple.

In such a rare scenario (which was exposed in some of the GR unit tests), flip
the peer gracefully.

It is to be noted in the normal GR state machine, most of the time is spent
in GR_TIMER_WAIT/LLGR_TIMER_WAIT states, which in the order of several minutes.
Hence it is not a common occurance for a peer session to come up at exactly the
same time as when the GR timers expire.

Tests marked as flaky still need to be monitored before unmarking some of them
if not all..

Change-Id: I497a3e84ed6222ae5cfc1cb80679c0a962e0ead9
Partial-Bug: 1733446