Schema changes for Type-5 EVPN
To handle EVPN Type-5 prefixes, all L2 virtual-networks are to be associated
with a single L3VRF virtual-network. This is used to route packets for all
children L2 virtual-networks.

There already exists a link between logical-router and virtual-network. In this
schema change, this link is annotated with attributes so that the relationship
can be futher qualified. This attribute defaults to existing behavior as
ExternalGateway. Additional L2VN and L3VN types are also added to the link
attribute.

Change-Id: I560596afb5f16e77c48b1ae926e608afd5c8c504
Implements: blueprint https://blueprints.launchpad.net/opencontrail/+spec/evpn-type-5
Depends-On: I76783f4a121a4e60a5f77735df281b6c7988840c
Partial-bug: 1636654