when the project is updated with quota limit for a
specific resource, quota_counter dictionary is updated
with the new resource quota counter only in the api-server
which served the project updated request. Which results in
KeyError when the next resource create is served by the other
api-servers.

To fix this, init quota counter in stateful create/update
functions with the resource quota limit.

Change-Id: Ifd5f389bd11e3a432a0d874ef11856bebecc7e22
Closes-Bug: 1745511