zk counters max_count value is reset
when the project is updated with quota limit for a
specific resource, quota counter's max_count is updated
with the new resource quota limt only in the api-server
which served the project update request. Which results in
OverQuota when the next resource create is served by the other
api-servers.

To fix this, init quota counter's max_count is modified
to new quota limt of that particular resource before
quota limt verification.

Change-Id: I393f683ee0a4000102e1ad9589813c36bb2d1133
Closes-Bug: 1745101