Delete resource is served by the api server
which is not intialized with resource quota counter object
In such case if quota limit is finite, should initialize counter
object with current resource count

When quota limit is updated from finite value to unlimited
for a resource, resource counter node in zookeeper will be
deleted. So if delete resource is served by the api server with
resource quota counter object and if quota is unlimited we
should remove the resource quota counter object from the
quota counters cache.

Closes-Bug: 1745665

Change-Id: I0a90a5875281d3a10d35666e45d157846baad870