contrail-api systemd template unit support
This requires changes to:
- contrail-status to be able to check workers
- systemd service unit template

Template instance string (part between the "@" character and the
suffix) should be a single digit and it works as worker id.

Warning: using multiple digit number as template instance string in
current setup would lead to "listening" on ports higher than 65536
and it would not work.

Other changes to contrail-api.service files:
- adding /var/lib/contrail to RW Directories to allow interaction
  with `keystone-signing` directory
- RestartSec=3s: by default systemd will stop restarting service if it
  fails to restart more than five times in ten seconds. To avoid This
  rate limit we increase the time between restart to 3s from 100ms.

Closes-Bug: #1748445

Change-Id: If203039f2bec147fe1823fe0b5e8125970f43925