when the project is updated with quota limit for a
specific resource, quota_counter dictionary is updated
with the new resource quota counter only in the api-server
which served the project updated request. Which results in
KeyError when the next resource create is served by the other
api-servers.

To fix this, init quota counter in stateful create/update
functions with the resource quota limit.

Closes-Bug: 1745511

Conflicts:
	src/config/api-server/vnc_cfg_api_server/tests/test_quota.py

Change-Id: Ib6813b431b92a717f76cf9db8bfa741d21572b57