Use bgp listen port number as configured
In current code, control-node bgp always listened to port based on command-line
or .conf based value (default 179). In this change, modify the same to use what
is configured in the schema database. Hence, bgp listen port option in the
contrail-control.conf file is deprecated.

Also, upon any change to listen port configuration or bind failure, restart the
process by sending SIGTERM to self (control-node) process. This is in line with
current behavior.

Change-Id: I9b9c5b380d564590c21f6c42c30a44e3065a2099
Partial-Bug: #1747826