WIP: Inter AS Option C - Part 2
Design details are in:
https://github.com/Juniper/contrail-specs/blob/master/5.0/inter_as_opt_c.md
The changes herein are to support a Tunnel Encapsulation List per
peer, per address family that will be used to influence the
encapsulation to use for routes received from a BGP peer that omits
the same.
The same Tunnel Encapsulation List will also be used to modify the
tunnel encapsulation for routes advertised to the peer, in this case
overriding any tunnel encapsulation present.
Details:
o BGP Schema changes to configure the Tunnel Encapsulation List per
  address family, per peer. Note that this list is not expected to be
  configured for non-VPN address families including labeled inet
  (earlier commit for supporting this feature). Added 'mpls' as an
  encapsulation supported in the XMPP schema, have NOT done the same
  for the multicast and Ethernet at this time.
o Apply the Tunnel Encapsulation List while inserting a path if
  configured for the family at hand on the peer the path is learned
  from. This is handled in ProcessRoutingPolicy.
o Insert the Tunnel Encapsulation List in the RibExportPolicy to ensure
  that the same Ribout is used for all peers configured with the same
  Tunnel Encapsulation List.
o in BgpTable::GetUpdateInfo, for BGP peers, update the outgoing Update
  with the configured Tunnel Encapsulation List if configured on the
  peer for the family at hand.

Pending:
Unit tests
Partial-Bug: #1743517
Change-Id: Ibd893ea2571d6a21699007dbed211d14b48494fc