Inter AS Option C - Part 2
Design details are in:
https://github.com/Juniper/contrail-specs/blob/master/5.0/inter_as_opt_c.md
The changes herein are to support an ordered Tunnel Encapsulation
List per peer, per address family that will be used to influence the
encapsulation to use for routes received from a BGP peer that omits
the same.
The same Tunnel Encapsulation List will also be used to modify the
tunnel encapsulation for routes advertised to the peer, in this case
overriding any tunnel encapsulation present.
Details:
o Apply the Tunnel Encapsulation List while inserting a path if
  configured for the family at hand. This is handled in
  ProcessRoutingPolicy for tables that support routing policies, labeled
  inet could be the only case at this time that supports routing policies
  and tunnel encapsulation lists. However, a minor additional change is
  required to not blindly strip communities.
  Processing for Default Tunnel Encapsulation on tables that do not
  support routing policies are handled directly in InsertPath.

o Insert the Tunnel Encapsulation List in the RibExportPolicy to ensure
  that the same Ribout is used for all peers configured with the same
  ordered Default Tunnel Encapsulation List.

o in BgpTable::GetUpdateInfo, for BGP peers, update the outgoing Update
  with the configured Tunnel Encapsulation List if configured on the
  peer for the family at hand.

o Show changes to display the configured Default Tunnel Encapsulation
  List.

o Integration tests added to verify functionality on received routes
  and routes advertised, including routes sent to XMPP peers.
  Verifies with and without tunnel encapsulations in the routes
  received as well as the changes to the routes when the Default
  Tunnel Encapsulation List is either modified or removed.

Partial-Bug: #1743517
Depends-On: I315b7290acbb72adea98546fb821ea4f738dd0be
Change-Id: Ibd893ea2571d6a21699007dbed211d14b48494fc