Add version agnostic targets for python packages
Instead of trying to figure out the correct version of vnc_api and
hardcoding that in other targets, create a generic target for the
library.

Do the same for cfgm_common and sandesh-common, even though both still
hardcode versions, to make rules.py changes consistent.

This is part of a patch series rewriting cfgm_common to tox:

  1. create version agnostic python targets <-- here
  2. transition rules.py to use generic targets
  3. rewrite cfgm_common to use tox
  4. update other SetupPyTestSuiteWithDeps calls
  5. get rid of rules.py transition

Partial-Bug: #1751438
Change-Id: I719a9c978e06aea464492d2d6e375e787eac1ff1