Add version agnostic target for
Instead of trying to figure out the correct version of vnc_api and
hardcoding that in other targets, create a generic target for the
library.

Do the same for cfgm_common (which does no versioning), so it can
also be used in rules.py.

This is part of a patch series rewriting cfgm_common to tox:

  1. create version agnostic python targets <--
  2. transition rules.py to use generic targets
  3. rewrite cfgm_common to use tox
  4. update other SetupPyTestSuiteWithDeps calls
  5. get rid of rules.py transition

Partial-Bug: #1751438
Change-Id: I719a9c978e06aea464492d2d6e375e787eac1ff1