Logical Interface should carry Physical Interface FQ name
If a corresponding physical interface does not exist, the
logical interface create fails.
Fixed the test to match the FQ names of physical interface
with logical interface.

Also catch an exception if such a failure happens.

The test fix is removed from this patch and is pushed under following
patch - https://review.opencontrail.org/#/c/40181/

Change-Id: Ib313429ae4ac1b54307a5cbfa0aa0a81252e139f
Partial-Bug: 1683547