Attach OriginVN extended-community to IP fabric routes
When agent re-advertises VPN routes into ip-fabric default inet.0 table,
VPN information is completely lost. This makes operations and debugging
very hard. Implementing routing-policies are also quite difficult as the
routes are not colored with any VPN information.

Agent should advertise primary instance index along with the NLRI when
advertising fabric routes.

Based on this index, control-node shall calculate appropriate source-rd
and store it as SourceRD attribute. Based on this source-rd, any inetvpn
route can be looked up on in bgp.l3vpn.0 table. If found and if this route
has any OriginVN extended-community associated with it, then that community
is also attached to the inet route. This attribute is then advertised to
other bgp peers as well.

Both inet and inetpvn routes are handled in the same DB partition as the
hashing is based only on the inet address. Hence, inetvpn table can be
safely accesed when processing inet routes and vice versa.

TODO: Agents need to tag ip-fabric routes with primary routing-instance index

Change-Id: Ic9203926416f2297ca0c2632ec31aaf68fe2c3f0
Partial-Bug: 1753397