Rearrange namespace label update order.
Contrail Config has the behavior that, unset tags on an object
is done only by key. The value of the tag is not taken into
consideration.

When value of a label is changed for a namespace, we need to first
remove the existing label and then add the new label with latest value.

Change-Id: I1b4fb8e7ca0ab8e757f5070d4702f323de6988d0
Closes-Bug: #1750217