CN fix to restart stale timer for initial configur
On IFMAPServer restart, when the control node re-connects to a different
IFMAPServer and receives the initial searchResult a  stale cleanup timer is
started, the timer is restarted for each subsequent searchResult.
Sometimes we notice that the IFMAPServer sends the initial configuration as
updateResults which do not restart the timer and  at times, all configiration is
not received by the control node when the timer expires.
This fix is to extend the timer by a minute(fudged) everytime we receive a
search update if the stale timer is running. If the update result is received
within the original timeout(minus fudge) we keep it unchanged.
Similar extention is applied to the end of RIB timer as well.
Partial-Bug: #1733478

Change-Id: Ida41ab4b2eb73b7cdef76672ac1adbe19a5f0c96