Unnecessary sessions getting logged with out SLO rate check
Idle sessions for which stats are not changed can be logged with SLO rate match
is not a bug.

This fix has below change:
With this change, all tear down sessions will not be logged.
Tear down sessions which are logged earlier at least once will be logged immidiately
with out checking rate, otherwise it will be considered as normal session and will be
reported only after matching the SLO rate.

Change-Id: I0f1ae5e4239fec9eebe57c8b34b6f20c3431403a
Closes-Bug: 1744914