Fixed lbaas-healthmonitor traceback issue
healthmonitor would be associated to pool in lbaas-v2.
so provider information would be taken from pool to
find the lbaas driver. If there is a crash in adding
the pool, the provider wont be set correctly in the pool
which ends storing healthmonitor object without provider.

Added a condition to check for the provider before calling
the health monitor object

Change-Id: I517050d8af90876e52e0ff88e2ef64c02834ce8b
Closes-bug: #1756056