[config] Change security draft mode model
Replace the boolean property 'pending_update' of a security resource by an enum
property to specified if it's a draft resource creation, modification or
deletion. If not set that means it's not a draft security resource.
Also don't unset draft mode on the scope after a commit and a revert
action on it.

Change-Id: I2c7dc31e99e3b4a690870b065f9969dff73cedff
Partial-Bug: #1742774
Depends-On: I8050c97f23cc74600ee659614d9f02a0e9a17561