Remove invalid test in XmppStateMachineTest
In the test Established_EvTcpClose_Then_EvXmppMessageStreamHeader,
EvXmppMessageStreamHeader is injected into a session_ XmppSession *, after
injecting EvTcpClose event. This is an invalid scenario because as a result of
processing EvTcpClose, XmppSession *session pointer can/will get destroyed.
Hence, we should not enqueue EvXmppMessageStreamHeader over a deleted session
later on. This can cause the test to crash and hence makes this test case in
xmpp_server_sm_test flaky.

Change-Id: I66b257c49ede25e4142d93aa0264f729b1896031
Closes-Bug: 1758806