[config] Catch RefExists exception on Tag deletion
All Tag resource have a reference to it's TagType resource. When a Tag
is deleted, code tries to delete its refered TagType and if that TagType
is still used by another Tag, it ignores TagType deletion. That patch
correctly handles that case to prevent stack trace in API server log.

Change-Id: I412c8b52a5c99b4e06733fa34cd10a439a41a694