peering control nodes in subcluster
Description: contrail control nodes of a sub cluster when provisioned
parallely, won't peer among themselves. This is because load/store is
not atomic.
In the current deployer, where nodes come up in parallel and provision
themselves in parallel, this is an issue. New deployer doesn't have
the option of provisioning it sequentially.
This script may be invoked once at the end of provisioining to peer
control nodes among themselves in all respective sub cluster.
It may also be invoked to peer control nodes of a specific sub cluster.
Marking it partial bug, as deployer changes are needed to call at
the end of provisioning.

Change-Id: I6da416b20807484577aae32eea5a0858ff5b3825
Partial-Bug: #1758191