Assert that deleted path attribute was indeed erased from the holder set
Ensure that when ever a path attribute is deleted, it is indeed correctly
erased from the set it is stored in by asserting on the return value
to the std::set::erase() call.

A hard bug was in bgp_server_test, where in a path attribute was injected
with the wrong path attribute DB, which later resulted in data corruption.

Eventually, I found out this root cause by adding this assert. But this
assert exposed tests which added path attributes directly to the db table
queue without adding them first to the attribute db. Fixes here ensure that
all those tests are also fixed

Change-Id: I91d65070a5279c3909b99132bb2f8010e185a272
Closes-Bug: 1759477