[config] Set draft switch property under Global System Config
In the first implementation of the draft mode for security resources, we
set the property switch under the project resource for the project scope
case and under the global policy management for global scope case. This
is not compliant which what we do habitually, a deployment config
property should be a property of the Global System Config resource.

Depends-On: Id1ef5769ae09f7a109f76362fa4bd2293f0626ba
Change-Id: I958e01061eb66c300de2a22dd4d729e82ad26fa5
Closes-Bug: #1759501