Policy match was not matching subprotocol correctly
If matching protocol is subprotocol, then match function should compare
it with corresponding value in attributes. Otherwise, match with path
protocol as it was done earlier.
Also need to retain subprotocol value in re-originated routes since
there can be a use case to match on subprotocol for reoriginated
routes. Will have to modify spec accordingly.

Change-Id: Idbbf3e216dfa6cf2fd0e298a451125750613ab83
Closes-Bug: #1756498