Sanity: mismatch in vrouter and agent nh in route
Ksync was not using new field rewrite_dmac as key while oper was doing
same. This used to result in inconsistency where oper used to point to a
NH which ksync never added. Vrouter was reflecting state of ksync and
used to drop packets because in flow source NH was not known to it. There
can be other manifestation of this problem as well.

Change-Id: I1890eed2f22ee610f0b2cd0c7e8752dcaf06c897
Closes-bug: #1761157