Adding debugs to graceful_restart_test suite
Tried running graceful_restart_flap_some_test2 a bunch of times
but did not see it fail.
Adding a few debugs to the graceful_restart_test suite so that
we can debug the failure better the next time we hit the issue.
UT'ed on local ws to make sure debugs are printed.

Partial-Bug: #1761337
Change-Id: I107c632d3a883989b8cb5dcbdb0fa4f147739770