Make Stale or EOR timeout increment configurable
If objects are created and deleted to verify sanity perioidically the stale or
end of RIB timers may never expire, currently that could happen if the periodic
create/delete operations are done at a frequency of a minute or less given the
increment is fixed at 60s.
Making the increment configurable allows the user to set is based on the
expected quiet period.
Changed the default value of the stale or EOR  timeout increment to 45s from 60s
Closes-Bug: 1761927
Change-Id: I5bae0c68206c9dbcb4a1d2245f42b5047a69c657