Sending Type-4 routes to agent for imported routes
Type-4 route is only generated if there is a global erpmvpn route. This
is generated only in one of the control-nodes for a given <S,G>. So for
receivers in other control-node, we do not end up sending type-4 route.

This case gets covered if we also send (over xmpp) any replicated type-4
route if the primary table of the type-4 route is bgp.mvpn.0
Partial-Bug: #1759753

Change-Id: I0518ebcaf47a7af1706ef36d9f2b705e37404724