Fix issue of multiple flow creation.
When fat-flow is configured we see different set of flows getting created for forward and reverse traffic. This is fixed by doing lookup for fat-flow on egress interface and generating key for reverse flow accordingly. Also, when same port number is used as source and destination port numbers, always mask source port for both forward and reverse flows.

Change-Id: I28a43cca04e0c15504446b4411704bed24076487
Closes-Bug: #1761517