Avoided fat flow creation for k8s haproxy loadbalancers
For k8s ingress, the backend is always a k8s-service which is
implemented as native loadbalancer and it is a fip.
When fat flow is enabled, if the end point is in the same
compute, the reverse flow wont be matched. Due to this
new flow/reverse-flow is created which sends the packet
to haproxy directly without doing snat and connection is
reset by haproxy.

So we are removing the fat-flow option for k8s-ingress.

Closes-bug: #1762383

Change-Id: I6265547eb380f12ca92714447d5c40ae5ecdcc42