db_manage v1.2: Re-work route target cleaning
Re-define the way the script recovers the route target inconsistencies.
The source of trust move from the schema transformer DB to the
zookeeper DB. If Config or Schema DBs contains a RT not locked in
zookeeper, it removes it. It does not heal missing RT in Config and
Schema DBs, it lets the schema transformer takes care of that when it'll
be re-initialized. It also cleans stale lock in zookeeper.

Change-Id: I2b61d446bc6baf4c384eea8504b2fa8999646518
Closes-Bug: #1764770
(cherry picked from commit 20fda499863151718d8f528aaafc8e0dabfb028a)