db_manage v1.2: Re-work route target cleaning
Re-define the way the script recovers the route target inconsistencies.
The source of trust move from the schema transformer DB to the
zookeeper DB. If Config or Schema DBs contains a RT not locked in
zookeeper, it removes it. It does not heal missing RT in Config and
Schema DBs, it lets the schema transformer takes care of that when it'll
be re-initialized. If missing RT persists in Config or/and in
Schema DB (foundable with command 'check_route_targets_id') that means
zookeeper have a stale lock which needs to be manually deleted.

Change-Id: I2b61d446bc6baf4c384eea8504b2fa8999646518
Closes-Bug: #1764770