Expanad tabs and remove trailing spaces
In a recent code review, unmodified code had trailing spaces which
always show up in "Red" in git diff and in gerrit. Instead of a point
fix, clean the entire C++ code base to expand tabs to spaces and also
remove all trailing spaces.

This may parts of code difficult for back-porting to other branches but
this code hygiene has been pending for a while. Hence we may only back
port to R5.0.

Change-Id: I24fec35e5c84de9819ab83a14546b0fa22b42fad
Partial-Bug: 1765458