DeleteUVEs generate core dump.
When we do delete UVEs, try to connect redis again. But connect failed.
From the code, we check ansync connection state, but from gdb, when redis
down, the ansync connection is not notified immedietly, that why problem
can't catch every time.

When redis connection is lost, we do coredump is not helpful. Just like
check ansync state, we simply exit contrail-collecor with error.

Change-Id: I1886ee799705b98c354f7bc145c188fdc01f377a
closes-bug: 1730786