[config] Fix race in security resource modification when draft mode enabled
In security draft mode to avoid issues, a scope lock is created when
an action is ongoing. It avoids any other action to be performed and also
forbidden any security resource modifications.
The check on resource modification was buggy as it required the lock to
check if it was already acquired. That produces race when concurrency
security resources modifications occur.

Change-Id: I980456c22551ca6e7f5dd645140f4f54fb354c0a
Closes-Bug: #1767425
(cherry picked from commit d60145153b1381d24ef81f809147a766bce6b9b2)