Read bgp_identifier_ in a thread-safe manner in HasSelfConfiguration()
During periodic uve send about the status of control-node as part of
the process state, BgpServer::HasSelfConfiguration() is invoked from
the main thread. This accesses boost::asio::ip::address_v4
bgp_identifier_, which is not thread-safe. Hence maintain a uint32_t
form of bgp_identifier_ and use it instead.

This was noticed while fixing bug 1766580.

Change-Id: Ie73becf7a58fd888a35601d9f925265fc48a2c3a
Closes-Bug: 1766580